moved App.jsx and started on React Router refactor
