+++ client/src/index.jsx
@@ -1,9 +1,27 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./components/App";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import App from "./App";
+import Dashboard from "./pages/Dashboard";
+
+const router = createBrowserRouter([
+    {
+      path: "/",
+      element: <App />,
+    //TODO: errorElement: <ErrorPage />,
+    },
+    {
+      path: "/dashboard",
+      element: <Dashboard />,
+    },
+    {
+      path: "/login",
+      element: <Dashboard />,
+    },
+  ]);
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <App />
+    <RouterProvider router={router} />
   </React.StrictMode>
 );
